Implement log refactor hypermedia
