+++ src/routes/log.routes.js
@@ -1,10 +1,10 @@
 const logService = require('../services/log.service')()
 const logController = require('../controllers/log.controller')(logService)
-const logHypermedia = require('../util/hypermedia/log.hypermedia')()
+const hypermedia = require('../util/hypermedia/hypermedia.config')()
 const routes = require('./routes.config')
 
 module.exports = router => {
-  router.get(routes.get('getLogs').path, logController.getAllLogs, logHypermedia.setResponse)
-  router.get(routes.get('getByLogCode').path, logController.getByCode, logHypermedia.setResponse)
-  router.get(routes.get('getByLogStatus').path, logController.getByStatus, logHypermedia.setResponse)
+  router.get(routes.get('getLogs').path, logController.getAllLogs, hypermedia.setResponse)
+  router.get(routes.get('getByLogCode').path, logController.getByCode, hypermedia.setResponse)
+  router.get(routes.get('getByLogStatus').path, logController.getByStatus, hypermedia.setResponse)
 }
